Make public API a bit more restricted
Instead of adding a new `WithStringRepresentation` public option, we make the `NewRetrievedFromYAML` method public.
This makes it so that people can't add arbitrary string representations for their retrieved values, but instead only if the value is provided as YAML.
