+++ confmap/provider/internal/configurablehttpprovider/provider.go
@@ -15,7 +15,6 @@ import (
 	"strings"
 
 	"go.opentelemetry.io/collector/confmap"
-	"go.opentelemetry.io/collector/confmap/provider/internal"
 )
 
 type SchemeType string
@@ -109,7 +108,7 @@ func (fmp *provider) Retrieve(_ context.Context, uri string, _ confmap.WatcherFu
 		return nil, fmt.Errorf("fail to read the response body from uri %q: %w", uri, err)
 	}
 
-	return internal.NewRetrievedFromYAML(body)
+	return confmap.NewRetrievedFromYAML(body)
 }
 
 func (fmp *provider) Scheme() string {
